+++ src/compas_timber/connections/t_lap.py
@@ -1,6 +1,6 @@
 from compas_timber.errors import BeamJoiningError
-from compas_timber.fabrication import JackRafterCut
-from compas_timber.fabrication import Lap
+from compas_timber.fabrication import JackRafterCutProxy
+from compas_timber.fabrication import LapProxy
 
 from .lap_joint import LapJoint
 from .solver import JointTopology
@@ -93,11 +93,11 @@ class TLapJoint(LapJoint):
 
         # create lap features
         negative_volume_main, negative_volume_cross = self._create_negative_volumes(self.cut_plane_bias)
-        main_lap_feature = Lap.from_volume_and_beam(negative_volume_main, self.main_beam, ref_side_index=self.main_ref_side_index)
-        cross_lap_feature = Lap.from_volume_and_beam(negative_volume_cross, self.cross_beam, ref_side_index=self.cross_ref_side_index)
+        main_lap_feature = LapProxy.from_volume_and_beam(negative_volume_main, self.main_beam, ref_side_index=self.main_ref_side_index)
+        cross_lap_feature = LapProxy.from_volume_and_beam(negative_volume_cross, self.cross_beam, ref_side_index=self.cross_ref_side_index)
 
         # cutoff feature for main beam
-        main_cut_feature = JackRafterCut.from_plane_and_beam(self.main_cutting_plane, self.main_beam)
+        main_cut_feature = JackRafterCutProxy.from_plane_and_beam(self.main_cutting_plane, self.main_beam)
 
         # add processings to the beams
         self.cross_beam.add_feature(cross_lap_feature)